+++ README.md
@@ -1,1 +1,12 @@
 # movies-explorer-api
+Репозиторий для  дипломного проекта, включающий  бэкенд часть приложения со следующими возможностями: авторизации и регистрации пользователей, операции с карточками и пользователями.  
+
+Адрес репозитория: https://oksana-bykova.github.io/movies-explorer-api/
+
+## Ссылки на проект
+
+IP 130.193.42.120
+
+Frontend https://oksana.bikova.movies.nomoredomains.xyz/
+
+Backend https://api.oksana.bikova.movies.nomoredomains.xyz/
