+++ routes/movies.js
@@ -1,41 +1,19 @@
 const router = require('express').Router();
 
-const { celebrate, Joi } = require('celebrate');
+// const { celebrate, Joi } = require('celebrate');
 const {
   getMovies,
   createMovie,
   deleteMovie,
 } = require('../controllers/movies');
 
-const { RegURL } = require('../utils/constants');
-
-const idValid = celebrate({
-  params: Joi.object().keys({
-    moviesId: Joi.string().length(24).hex().required(),
-  }),
-});
+const { validateCreateMovie, idValid } = require('../middlwares/validation');
 
 router.get('/movies', getMovies);
 
 router.post(
   '/movies',
-  celebrate({
-    body: Joi.object().keys({
-      name: Joi.string().required().min(2).max(30),
-      link: Joi.string().required().regex(RegURL),
-      country: Joi.string().required(),
-      director: Joi.string().required(),
-      duration: Joi.number().required(),
-      year: Joi.string().required(),
-      description: Joi.string().required(),
-      image: Joi.string().required().regex(RegURL),
-      trailer: Joi.string().required().regex(RegURL),
-      nameRU: Joi.string().required(),
-      nameEN: Joi.string().required(),
-      thumbnail: Joi.string().required().regex(RegURL),
-      movieId: Joi.number().required(),
-    }),
-  }),
+  validateCreateMovie,
   createMovie,
 );
