+++ routes/users.js
@@ -1,10 +1,12 @@
 const router = require('express').Router();
-const { celebrate, Joi } = require('celebrate');
+
 const {
   getUsers,
   updateProfile,
   getUserMe,
 } = require('../controllers/users');
+
+const { validateUpdateProfile } = require('../middlwares/validation');
 
 router.get('/users', getUsers);
 
@@ -12,12 +14,7 @@ router.get('/users/me', getUserMe);
 
 router.patch(
   '/users/me',
-  celebrate({
-    body: Joi.object().keys({
-      name: Joi.string().min(2).max(30),
-      about: Joi.string().min(3).max(30),
-    }),
-  }),
+  validateUpdateProfile,
   updateProfile,
 );
